Feature: Users don't have to add locations anymore
Through /test if a location is searched then it is added to the
database. TODO: Check if location is already in the database, before
adding it. TODO: Specify Location object KEY to be the same as Gooogle
PlacesID, that way we can use get_by_id which is faster than
query(googlePlaceId = request.self.get("PlaceId"))